# InfinityRunWEBAPI/InfinityRunWEBAPI/Controllers/UserController.cs
using InfinityRunWEBAPI.Models;
using InfinityRunWEBAPI.Services;
using Microsoft.AspNetCore.Mvc;
using System.Text;
using System.Security.Cryptography;

namespace InfinityRunWEBAPI.Controllers;

[ApiController]
[Route("api/[controller]")]
public class UserController : ControllerBase
{
    private readonly UserService _usersService;

    public UserController(UserService usersService) =>
        _usersService = usersService;

    [HttpGet]
    public async Task<List<User>> Get() =>
        await _usersService.GetAsync();

    [HttpGet("{id:length(24)}")]
    public async Task<ActionResult<User>> Get(string id)
    {
        var user = await _usersService.GetAsync(id);

        if (user is null)
        {
            return NotFound();
        }

        return user;
    }

    [HttpPost]
    public async Task<IActionResult> Post(User newUser)
    {
        string data = newUser.password ?? string.Empty;
        string result = string.Empty;

        using (var myHash = SHA256.Create())
        {

            var byteArrayResultOfRawData =
                  Encoding.UTF8.GetBytes(data);


            var byteArrayResult =
                 myHash.ComputeHash(byteArrayResultOfRawData);


            result =
              string.Concat(Array.ConvertAll(byteArrayResult,
                                   h => h.ToString("X2")));
        }

        newUser.password = result;

        await _usersService.CreateAsync(newUser);

        return CreatedAtAction(nameof(Get), new { id = newUser.Id }, newUser);
    }

    [HttpPut("{id:length(24)}")]
    public async Task<IActionResult> Update(string id, User updatedUser)
    {
        var user = await _usersService.GetAsync(id);

        if (user is null)
        {
            return NotFound();
        }

        updatedUser.Id = user.Id;

        await _usersService.UpdateAsync(id, updatedUser);

        return NoContent();
    }

    [HttpDelete("{id:length(24)}")]
    public async Task<IActionResult> Delete(string id)
    {
        var user = await _usersService.GetAsync(id);

        if (user is null)
        {
            return NotFound();
        }

        await _usersService.RemoveAsync(id);

        return NoContent();
    }
}